requirements.txt
praw
requests
pyfcm
python-firebase
flask
html2text
feedparser
beautifulsoup4
gcloud
